Reorganize unit tests and add e2e playwright tests
